Compile styles to a stylesheet on demand only
